+++ src/amun/common/beam.h
@@ -11,8 +11,8 @@ namespace amunmt {
 typedef std::vector<HypothesisPtr> Beam;
 typedef std::vector<Beam> Beams;
 
-std::string Debug(const Beam &vec, size_t verbosity = 1);
-std::string Debug(const Beams &vec, size_t verbosity = 1);
+std::string Debug(const Beam &vec, unsigned verbosity = 1);
+std::string Debug(const Beams &vec, unsigned verbosity = 1);
 
 }
